metta_repl

Module metta_shim

MettaShim is responsible for ALL calls between the repl and MeTTa, and is in charge of keeping Python happy (and perhaps other languages in the future).

Because so much functionality can be extended with Python, MettaShim must handle many operations you might not expect, such as rendering atoms to text or even dropping atoms

Re-exports

pub use metta_interface_mod::MettaShim;

Modules

metta_interface_mod

The “no python” path involves a reimplementation of all of the MeTTa interface points calling MeTTa directly instead of through Python. Maintaining two paths is a temporary stop-gap solution because we can only link the Hyperon Rust library through one pathway and the HyperonPy module is that path when the Python repl is used.

Functions

exec_state_prepare

Prepares to enter an interruptible exec loop

exec_state_should_break

Check whether an exec loop should break based on an interrupt, and clear the interrupt state

parse_and_unroll_syntax_tree